chore: move some functionality from Config to ConfigModule
