+++ src/config/config.rs
@@ -14,10 +14,10 @@ use crate::config::from_document::from_document;
 use crate::config::source::Source;
 use crate::directive::DirectiveCodec;
 use crate::http::Method;
+use crate::is_default;
 use crate::json::JsonSchema;
 use crate::merge_right::MergeRight;
 use crate::valid::{Valid, Validator};
-use crate::{is_default, scalar};
 
 #[derive(
     Serialize, Deserialize, Clone, Debug, Default, Setters, PartialEq, Eq, schemars::JsonSchema,
@@ -67,27 +67,8 @@ impl Config {
         self.server.port.unwrap_or(8000)
     }
 
-    pub fn output_types(&self) -> HashSet<&String> {
-        let mut types = HashSet::new();
-        let input_types = self.input_types();
-
-        if let Some(ref query) = &self.schema.query {
-            types.insert(query);
-        }
-
-        if let Some(ref mutation) = &self.schema.mutation {
-            types.insert(mutation);
-        }
-        for (type_name, type_of) in self.types.iter() {
-            if (type_of.interface || !type_of.fields.is_empty())
-                && !input_types.contains(&type_name)
-            {
-                for (_, field) in type_of.fields.iter() {
-                    types.insert(&field.type_of);
-                }
-            }
-        }
-        types
+    pub fn find_type(&self, name: &str) -> Option<&Type> {
+        self.types.get(name)
     }
 
     pub fn recurse_type<'a>(&'a self, type_of: &str, types: &mut HashSet<&'a String>) {
@@ -99,37 +80,6 @@ impl Config {
                 }
             }
         }
-    }
-
-    pub fn input_types(&self) -> HashSet<&String> {
-        let mut types = HashSet::new();
-        for (_, type_of) in self.types.iter() {
-            if !type_of.interface {
-                for (_, field) in type_of.fields.iter() {
-                    for (_, arg) in field
-                        .args
-                        .iter()
-                        .filter(|(_, arg)| !scalar::is_scalar(&arg.type_of))
-                    {
-                        if let Some(t) = self.find_type(&arg.type_of) {
-                            t.fields.iter().for_each(|(_, f)| {
-                                types.insert(&f.type_of);
-                                self.recurse_type(&f.type_of, &mut types)
-                            })
-                        }
-                        types.insert(&arg.type_of);
-                    }
-                }
-            }
-        }
-        types
-    }
-    pub fn find_type(&self, name: &str) -> Option<&Type> {
-        self.types.get(name)
-    }
-
-    pub fn find_union(&self, name: &str) -> Option<&Union> {
-        self.unions.get(name)
     }
 
     pub fn to_yaml(&self) -> Result<String> {
@@ -165,10 +115,6 @@ impl Config {
         }
         self.types = graphql_types;
         self
-    }
-
-    pub fn contains(&self, name: &str) -> bool {
-        self.types.contains_key(name) || self.unions.contains_key(name)
     }
 }
