+++ src/config/into_document.rs
@@ -2,14 +2,14 @@ use async_graphql::parser::types::*;
 use async_graphql::{Pos, Positioned};
 use async_graphql_value::{ConstValue, Name};
 
-use super::Config;
+use super::{Config, ConfigModule};
 use crate::blueprint::TypeLike;
 use crate::directive::DirectiveCodec;
 
 fn pos<A>(a: A) -> Positioned<A> {
     Positioned::new(a, Pos::default())
 }
-fn config_document(config: &Config) -> ServiceDocument {
+fn config_document(config: &ConfigModule) -> ServiceDocument {
     let mut definitions = Vec::new();
     let mut directives = vec![
         pos(config.server.to_directive()),
@@ -258,6 +258,6 @@ fn get_directives(field: &crate::config::Field) -> Vec<Positioned<ConstDirective
 
 impl From<Config> for ServiceDocument {
     fn from(value: Config) -> Self {
-        config_document(&value)
+        config_document(&value.into())
     }
 }
